Reorganize code and add localstorage
