--- client/src/store.js
@@ -1,4 +1,0 @@
-import { writable } from "svelte/store";
-
-// Defaults to the cinema_id of Pathé Delft (18)
-export const cinema_id = writable("18");
